requirements.txt
aiohttp
aria2p==0.9.1
beautifulsoup4
bs4
cfscrape
feedparser
hachoir
hurry.filesize
js2py
lk21
lxml
messages
Pillow
psutil
pyrogram==1.4.16
python-dotenv
pyprog
python-telegram-bot
pybase64
requests
speedtest-cli
tgcrypto
urllib3
yt-dlp
telegraph
mediainfo
